removed writeFile,js
