+++ src/utils.js
@@ -5,8 +5,46 @@ import uuidv1 from 'uuid/v1';
 import dayjs from 'dayjs';
 import _ from 'lodash';
 import path, { resolve } from 'path';
-import { read, dirSync, syncStats } from './fileSystem';
-import { stripSymbols, getFileName, updateContent } from './writeFile';
+import { readdirSync, statSync, readFile } from 'fs';
+import isValid from 'is-valid-path';
+
+/**
+ * read()
+ * @param {string} absolutePath
+ *
+ */
+const read = (absolutePath) => new Promise((res, rej) => {
+  console.log(absolutePath);
+  if (!isValid(absolutePath)) {
+    console.log('path is invalid');
+  }
+  let dataStr;
+  readFile(absolutePath, 'utf8', (err, data) => {
+    if (!err) {
+      if (data === '') {
+        console.log(`${absolutePath} returned empty`);
+      }
+      dataStr = JSON.parse(data);
+      console.log(dataStr);
+      res(dataStr);
+    } else {
+      console.log(err);
+      rej(err);
+    }
+  });
+});
+
+/**
+ * @param {string} path
+ *
+ */
+const dirSync = (filepath) => readdirSync(filepath);
+
+/**
+ * @param {string} path
+ *
+ */
+const syncStats = (filepath) => statSync(filepath);
 
 const checkFilePath = async (filePath) => {
   if (await pathExists(filePath)) {
@@ -155,6 +193,85 @@ const setupPath = (pathToSrc) => {
   return files;
 };
 
+// Added writeFile.js functions
+/**
+ * for stripSymbols()
+ * @param {Object} data a json object
+ *
+ */
+const stripSymbols = (data) => {
+  let dataStr = JSON.stringify(data);
+
+  const replaceList = [
+    ['/{"/g', '{ "'],
+    ['/{"/g', '{ " '],
+    ['/},{/g', ' },\n{'],
+    ['/":/g', '": '],
+    ['/,"/g', ',\n "'],
+  ];
+
+  replaceList.forEach((replacer) => {
+    dataStr = dataStr.replace(replacer[0], replacer[1]);
+  });
+
+  return dataStr;
+};
+
+/**
+ * fixFileName()
+ * @param {string} fileName
+ */
+const fixFileName = (fileName) => {
+  let correctedFileName;
+
+  correctedFileName = fileName.replace(/ /g, '_'); // Replace space with underscore
+  correctedFileName = fileName.toLowerCase(); // Maintain Uniformity
+
+  return correctedFileName;
+};
+
+/**
+ * getFileName()
+ * @param {string} file
+ * @param {Object} fileData
+ * @param {var} flag
+ * @param {var} index
+ */
+// @TODO if we use fileData.name - why we didn't just pass it here?
+const getFileName = (file, fileData, flag, index) => {
+  let fileName;
+  if (flag === 1) {
+    // for example: 23-someJsonFile.json
+    fileName = `${index}-${file}`;
+  } else {
+    // for example: someValueOfName.json
+    fileName = `${fileData.name}.json`;
+  }
+
+  fileName = fixFileName(fileName);
+  return fileName;
+};
+
+/**
+ * For updateContent()
+ * @param {var} content
+ * @param {var} keys
+ */
+const updateContent = (content, keys) => {
+  const contentCopy = content;
+
+  // @TODO error  Assignment to property of function parameter 'obj'
+  // no - param - reassign;
+  contentCopy.forEach((contentElem) => {
+    contentElem.forEach((obj) => {
+      keys.forEach((key) => {
+        delete obj[key];
+      });
+    });
+  });
+  return contentCopy;
+};
+
 export {
   checkFilePath,
   readAllFiles,
@@ -169,5 +286,6 @@ export {
   getFileKey,
   stripSymbols,
   getFileName,
+  fixFileName,
   updateContent,
 };
